move pure c++ code out of python string to a separate header file
